seperating out functionality
