+++ backend/components/database.py
@@ -1,0 +1,5 @@
+import components.embedding as embedding
+from langchain_chroma import Chroma
+
+def get():
+    return Chroma(persist_directory="./chroma_db1", embedding_function=embedding.get())
