+++ backend/components/embedding.py
@@ -1,0 +1,7 @@
+from langchain_ollama import OllamaEmbeddings
+
+def get():
+    return OllamaEmbeddings(
+        model="nomic-embed-text",
+        base_url="http://ollama:11434"
+    )
